Import json-rpc client library
