+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "async-jsonrpc-client"
+name = "jsonrpc-client"
 version = "0.1.3"
 authors = ["koushiro <koushiro.cqx@gmail.com>"]
 edition = "2018"
@@ -9,12 +9,6 @@ license = "MIT OR Apache-2.0"
 repository = "https://github.com/koushiro/async-jsonrpc-client-rs"
 keywords = ["async", "jsonrpc", "client"]
 
-[features]
-default = ["http", "ws"]
-http = ["reqwest"]
-ws = ["async-tungstenite", "parking_lot", "tokio"]
-http-surf = ["surf", "anyhow"]
-
 [dependencies]
 async-trait = "0.1"
 futures = "0.3"
@@ -22,18 +16,8 @@ log = "0.4"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 thiserror = "1.0"
-
-# HTTP
-reqwest = { version = "0.10", features = ["json"], optional = true }
-
-# HTTP (async-std)
-surf = { version = "2.1", optional = true, default-features = false, features = ["h1-client"] }
-anyhow = { version = "1", optional = true }
-
-# WebSocket
-async-tungstenite = { version = "0.8", features = ["tokio-runtime"], optional = true }
-parking_lot = { version = "0.11", optional = true }
-tokio = { version = "0.2", features = ["macros"], optional = true }
+surf = { version = "2.1", default-features = false, features = ["h1-client"] }
+anyhow = { version = "1" }
 
 [dev-dependencies]
 env_logger = "0.7"
