+++ src/errors.rs
@@ -11,10 +11,6 @@ pub enum RpcError {
     #[error("{0}")]
     RpcResponse(#[from] crate::types::Error),
     /// HTTP error.
-    #[cfg(feature = "http")]
-    #[error("{0}")]
-    Http(#[from] reqwest::Error),
-    #[cfg(feature = "http-surf")]
     /// HTTP error (surf).
     #[error("{0}")]
     Http(anyhow::Error),
